Add files via upload
The Air vehicle level Functional Hazard Analysis  (AFHA) identifies the Failure Conditions (FCs), their effects, and the severity classes of these effects. 
It is an input of many activities pertaining to the safety assessment process. In particular, it is requested for performing the Preliminary  Air vehicle level Safety Assessment (PASA).
The Failure Conditions are also requested when evaluating satisfiability of the Overarching Properties