Add files via upload
This document provides a selection of regulatory safety objectives from CS 25 Amendment 18 (2016), the European airworthiness regulation for large aéroplanes  (FAR 25 in the US).. 
It is the entry point to carry out the safety case on muXAV using the Overarching Properties and their satisfiability criteria.
The document also provides the Accepted Means of Compliance associated to the selected CS paragraphs. It is intended to provide material supporting assessment of the OPs and criteria as a candidate AMC against CS/FAR 25.1309. and AMC 25.1309.