Layer_2_MMS_SW_SPARK: use dimension system for floating-point units
Some units in types.ads are defined as floating-point types and others
as signed integer types. Start using the dimension system in GNAT on
floating-point units. Next step is to see if there is a benefit in using
floating-point types instead of signed integer types for the input and
output types, so that we can use the standard dimension system. Otherwise,
we could use a signed integer type with the Dimension_System aspect as
the root type for all these signed integer types.